+++ packages/smart_flutter_utils/lib/src/validators/verification_code.dart
@@ -4,7 +4,7 @@ enum VerificationCodeValidationError { empty, small }
 
 class VerificationCode
     extends FormsInput<String, VerificationCodeValidationError> {
-  const VerificationCode.pure() : super.pure('');
+  const VerificationCode.pure([super.value = '']) : super.pure();
   const VerificationCode.dirty([super.value = '']) : super.dirty();
 
   @override
@@ -17,4 +17,28 @@ class VerificationCode
       return null;
     }
   }
+}
+
+extension VerificationCodeFromString on String {
+  VerificationCode toVerificationCode() => VerificationCode.dirty(this);
+
+  VerificationCode toOptionalVerificationCode() {
+    final text = trim();
+    return text.isEmpty
+        ? text.toPureVerificationCode()
+        : text.toVerificationCode();
+  }
+
+  VerificationCode toPureVerificationCode() => VerificationCode.pure(this);
+}
+
+extension VerificationCodeFromStringNullable on String? {
+  VerificationCode toVerificationCode([String fallback = '']) =>
+      (this ?? fallback).toVerificationCode();
+
+  VerificationCode toOptionalVerificationCode([String fallback = '']) =>
+      (this ?? fallback).toOptionalVerificationCode();
+
+  VerificationCode toPureVerificationCode([String fallback = '']) =>
+      (this ?? fallback).toPureVerificationCode();
 }